requirements.txt
# PDF Processing
pdfplumber>=0.9.0

# Excel Processing
openpyxl>=3.1.0

# Configuration Management
PyYAML>=6.0
python-dotenv>=1.0.0

# Email Processing
# Note: imaplib, smtplib, email are built-in to Python

# Optional: Better fuzzy matching (can be added later)
# python-Levenshtein>=0.21.0
# fuzzywuzzy>=0.18.0
